requirements.txt
Django~=4.1
djangorestframework==3.14.0
setuptools==65.0.1
django-environ==0.9.0
django-cors-headers==3.13.0
djangorestframework-simplejwt==5.2.2

djangorestframework-api-key==2.2.0
drf-yasg==1.21.4
django-cron==0.6.0
coreapi~=2.3.3
requests==2.31.0

